Renamed folder name to refresh_data
